# tests/api/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "DiscardPreview": {
        "additionalProperties": false,
        "description": "What a discard would destroy, counted before anything is destroyed.\n\nOne shape for all three kinds (\u00a717.5), mirroring `DiscardPreviewView`\nfield-for-field plus its two computed properties \u2014 the wire format carries\nthem as data so a client does not re-derive a rule the service owns.",
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "active_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Detail"
          },
          "blocked": {
            "default": false,
            "title": "Blocked",
            "type": "boolean"
          },
          "cited_by": {
            "default": 0,
            "title": "Cited By",
            "type": "integer"
          },
          "extractions": {
            "title": "Extractions",
            "type": "integer"
          },
          "files": {
            "title": "Files",
            "type": "integer"
          },
          "has_exportable": {
            "default": false,
            "title": "Has Exportable",
            "type": "boolean"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "mismatches": {
            "title": "Mismatches",
            "type": "integer"
          },
          "runs": {
            "title": "Runs",
            "type": "integer"
          },
          "scores": {
            "title": "Scores",
            "type": "integer"
          },
          "tagged_mismatches": {
            "title": "Tagged Mismatches",
            "type": "integer"
          },
          "target_id": {
            "title": "Target Id",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "target_id",
          "label",
          "runs",
          "extractions",
          "scores",
          "mismatches",
          "tagged_mismatches",
          "files",
          "active"
        ],
        "title": "DiscardPreview",
        "type": "object"
      },
      "DiscardResponse": {
        "additionalProperties": false,
        "description": "What a discard actually removed.\n\nA body rather than a bare `204`, because the counts are the only record\nthat survives the call: nothing is written anywhere saying this happened\n(\u00a717.3), so the response *is* the receipt.",
        "properties": {
          "extractions": {
            "title": "Extractions",
            "type": "integer"
          },
          "files": {
            "title": "Files",
            "type": "integer"
          },
          "forced": {
            "default": false,
            "title": "Forced",
            "type": "boolean"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "mismatches": {
            "title": "Mismatches",
            "type": "integer"
          },
          "runs": {
            "title": "Runs",
            "type": "integer"
          },
          "scores": {
            "title": "Scores",
            "type": "integer"
          },
          "tagged_mismatches": {
            "title": "Tagged Mismatches",
            "type": "integer"
          },
          "target_id": {
            "title": "Target Id",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "target_id",
          "runs",
          "extractions",
          "scores",
          "mismatches",
          "tagged_mismatches",
          "files"
        ],
        "title": "DiscardResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "The body of every 4xx that is not a validation error.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PageMeta": {
        "additionalProperties": false,
        "description": "Paging envelope. `total` is the unpaged count matching the filter.",
        "properties": {
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "sort_dir": {
            "$ref": "#/components/schemas/SortDir"
          },
          "sort_key": {
            "title": "Sort Key",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "page",
          "page_size",
          "sort_key",
          "sort_dir"
        ],
        "title": "PageMeta",
        "type": "object"
      },
...
      "RunPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "items",
          "meta"
        ],
        "title": "RunPage",
        "type": "object"
      },
      "RunProgressResponse": {
        "additionalProperties": false,
        "description": "One per-model progress card. Every count is derived from committed\n`extraction` rows, never from a counter column (\u00a715 F6).",
        "properties": {
          "done": {
            "default": 0,
            "title": "Done",
            "type": "integer"
          },
          "elapsed_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Ms"
          },
          "eta_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eta Ms"
          },
          "median_latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Median Latency Ms"
          },
          "model_tag": {
            "title": "Model Tag",
            "type": "string"
          },
          "parse_failures": {
            "default": 0,
            "title": "Parse Failures",
            "type": "integer"
          },
          "percent": {
            "default": 0.0,
            "title": "Percent",
            "type": "number"
          },
          "prompt_tokens": {
            "default": 0,
            "title": "Prompt Tokens",
            "type": "integer"
          },
          "retries": {
            "default": 0,
            "title": "Retries",
            "type": "integer"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "total": {
            "default": 0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "run_id",
          "model_tag",
          "status"
        ],
        "title": "RunProgressResponse",
        "type": "object"
      },
      "RunResponse": {
        "additionalProperties": false,
        "description": "One row of the \"Runs in this evaluation\" table.",
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "evaluation_id": {
            "title": "Evaluation Id",
            "type": "string"
          },
          "is_dev": {
            "default": false,
            "title": "Is Dev",
            "type": "boolean"
          },
          "model_digest": {
            "title": "Model Digest",
            "type": "string"
          },
          "model_tag": {
            "title": "Model Tag",
            "type": "string"
          },
          "records_done": {
            "default": 0,
            "title": "Records Done",
            "type": "integer"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          }
        },
        "required": [
          "run_id",
          "evaluation_id",
          "model_tag",
          "model_digest",
          "status"
        ],
        "title": "RunResponse",
        "type": "object"
      },
      "RunStatus": {
        "description": "One model's pass over one evaluation (sw-design.md \u00a715.2).",
        "enum": [
          "queued",
          "running",
          "done",
          "failed",
          "interrupted"
        ],
        "title": "RunStatus",
        "type": "string"
      },
...
      "SortDir": {
        "description": "Sort direction. Always an explicit **service-call parameter**, even when\nthe dataset is small enough to sort in memory (sw-design.md \u00a78.4).",
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortDir",
        "type": "string"
      },
...
      "TaskAcceptedResponse": {
        "additionalProperties": false,
        "description": "Work was submitted; poll `GET /api/v1/tasks/{task_id}`.",
        "properties": {
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id"
        ],
        "title": "TaskAcceptedResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Road accident report analysis \u2014 import, census and evaluation.",
    "title": "RA2",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/runs": {
      "get": {
        "description": "The \"Runs in this evaluation\" table \u2014 paginated at 10, as drawn.",
        "operationId": "list_runs_api_v1_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "evaluation_id",
            "required": true,
            "schema": {
              "title": "Evaluation Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 10,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort_key",
            "required": false,
            "schema": {
              "default": "started_at",
              "title": "Sort Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort_dir",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortDir",
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}": {
      "delete": {
        "description": "Discard one run and, by the schema's cascade, everything derived from\nit (\u00a717.1).\n\n409 when the run is `queued` or `running` (G1), and when it carries tagged\nmismatches and `force` was not passed (G2). The second is overridable; the\nfirst is not.\n\nThe preview is read **first**, inside the same request, because after the\ndelete those counts do not exist anywhere \u2014 this response is the receipt.",
        "operationId": "discard_run_api_v1_runs__run_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "force",
            "required": false,
            "schema": {
              "default": false,
              "title": "Force",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscardResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discard Run",
        "tags": [
          "runs"
        ]
      },
      "get": {
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}/discard-preview": {
      "get": {
        "description": "What discarding this run would destroy. Counts only; writes nothing.",
        "operationId": "discard_preview_api_v1_runs__run_id__discard_preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscardPreview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discard Preview",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}/mismatches.csv": {
      "get": {
        "description": "The half a re-run cannot reproduce: `analyst_tag` and its note.",
        "operationId": "export_mismatches_api_v1_runs__run_id__mismatches_csv_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Mismatches",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}/progress": {
      "get": {
        "description": "Derived from committed `extraction` rows, never from a counter.",
        "operationId": "get_progress_api_v1_runs__run_id__progress_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunProgressResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Progress",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/v1/runs/{run_id}/resume": {
      "post": {
        "description": "Pick an interrupted run back up from its last committed extraction.",
        "operationId": "resume_api_v1_runs__run_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resume",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}/scores.csv": {
      "get": {
        "description": "Export before discard (\u00a717.3) \u2014 UTF-8 with BOM, `;`-delimited.",
        "operationId": "export_scores_api_v1_runs__run_id__scores_csv_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Scores",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}